feat(reactotron): Move android device help component out of the main Help page
